create file-based period model
