--- db/migrate/20160801080454_create_course_periods.rb
@@ -1,12 +1,0 @@
-class CreateCoursePeriods < ActiveRecord::Migration
-  def change
-    create_table :course_periods do |t|
-      t.string  :organization_code, null: false
-      t.string  :code
-      t.integer :order
-      t.string  :time
-
-      t.timestamps
-    end
-  end
-end
