+++ lib/course_period/errors.rb
@@ -1,0 +1,3 @@
+module CoursePeriod
+  class InvalidTimeFormat < StandardError; end
+end
