+++ lib/course_period/period.rb
@@ -1,0 +1,37 @@
+module CoursePeriod
+  class Period
+    attr_accessor :order, :code, :time, :start_at, :end_at
+
+    def initialize(row)
+      self.order, self.code, self.time = row
+      self.order = order.to_i
+
+      self.start_at, self.end_at = time.split('-').map { |s| Time.zone.parse(s) }
+
+      validate_time
+    end
+
+    def validate_time
+      # time string should match "10:10", two digits seperate by a colon
+      time_regex = /\d{2}\:\d{2}/
+      start_at_tmp, end_at_tmp = time.split('-')
+
+      raise InvalidTimeFormat unless time.include?('-')
+      raise InvalidTimeFormat unless start_at_tmp.match(time_regex) && end_at_tmp.match(time_regex)
+    end
+
+    def start_time
+      utc_format_str(start_at)
+    end
+
+    def end_time
+      utc_format_str(end_at)
+    end
+
+    private
+
+    def utc_format_str(time)
+      time.getutc.iso8601.gsub(/[-:]/, '')
+    end
+  end
+end
