--- spec/factories/course_periods.rb
@@ -1,5 +1,0 @@
-FactoryGirl.define do
-  factory :course_period do
-
-  end
-end
